+++ src/source/typecheck.rs
@@ -1054,6 +1054,7 @@ fn add_result(
     }
     match (left, right) {
         (Type::Num, Type::Num) => Ok(Type::Num),
+        (Type::Bytes, Type::Bytes) => Ok(Type::Bytes),
         (Type::List(left), Type::List(right)) => Ok(Type::List(match (left, right) {
             (Some(left), Some(right)) => Some(Box::new(Type::union([
                 left.as_ref().clone(),